Create PaymentMethod class
Fixes #84
